framed out Sidebar and homeview
I did a very initial framing out of the dashboard sidebar component named Header(we should change it to something better). also did an inital pass on the home view cards. Styling is going to be fun!

Also we are going to need to think about data flow to all these components. Also i think we are going to need to be moving all these icons around as links to assets right they are passing svg components, which is probably a terrible idea